feat(Calculator, AddCalculator, AddCalculatorTest): 서비스 코드 구현, 테스트 코드 작성

+++ src/main/java/calculator/service/Calculator.java
@@ -1,7 +1,5 @@
 package calculator.service;
 
-import java.util.List;
-
 public interface Calculator {
-    public Long calculate(String input);
+    public Integer calculate(String input);
 }
